+++ CHANGELOG.md
@@ -338,3 +338,4 @@
 - 2024-03-19: Metrics find email, discovery, lead, bettercontact, +2
 - 2024-03-20: Serde follow up, main, chat, README, +5
 - 2024-03-21: Err sender, icp, find email, campaigns, +5
+- 2024-03-22: Retry mod, users, mutate, descend, +6